+++ src/js/mixin/animate.js
@@ -79,14 +79,14 @@ export default {
 
             addClass(this.target, targetClass);
             children.forEach((el, i) => propsFrom[i] && css(el, propsFrom[i]));
-            css(this.target, 'minHeight', oldHeight);
+            css(this.target, 'height', oldHeight);
             window.scroll(window.pageXOffset, oldScrollY);
 
             return Promise.all(children.map((el, i) =>
                 propsFrom[i] && propsTo[i]
                     ? Transition.start(el, propsTo[i], this.animation, 'ease')
                     : Promise.resolve()
-            ).concat(Transition.start(this.target, {minHeight: newHeight}, this.animation, 'ease'))).then(() => {
+            ).concat(Transition.start(this.target, {height: newHeight}, this.animation, 'ease'))).then(() => {
                 children.forEach((el, i) => css(el, {display: propsTo[i].opacity === 0 ? 'none' : '', zIndex: ''}));
                 reset(this.target);
                 this.$update(this.target);
@@ -122,7 +122,7 @@ function reset(el) {
         width: ''
     });
     removeClass(el, targetClass);
-    css(el, 'minHeight', '');
+    css(el, 'height', '');
 }
 
 function getPositionWithMargin(el) {