+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Fixed
 
 - Fix active row style for striped table
+- Fix height of Masonry Grid
 
 ## 3.0.0 rc 5 (June 4, 2018)